Add files via upload
he Indian Premier League, or IPL, is a T20 cricket league, which was founded in 2008 and is held every year. It sees participation from both national and international players, and eight teams representing eight Indian cities compete with each other in a double round-robin format in the league stages, which are followed by playoffs. Over the years, IPL has become one of the most-watched and most attended live sporting events all over the world.
 
As a data analyst at IFP, a nationally recognized news agency, the Sports Editor of the agency has approached you to build an interactive Tableau dashboard of IPL statistics over the years since its inception in order to create an infographic for a newsletter that their team is working on.